[cluster] The plugin was removed and split into multiple plugins
The cluster plugin originally was used for all the components for clustering:
corosync, cman, pacemaker, gfs2, and dlm. This patch moves the command
output/file that was captured in the cluster plugin to the corresponding plugin
related to the command output or file. The cluster plugin was removed.

Each component that is used by the 2 different types of clusters have a
corresponding plugin: corosync, cman, pacemaker, gfs2, and dlm.

The 2 different types of a cluster are: cman (with or without rgmanager or
pacemaker managing resources) and pacemaker based clusters. Both of these
cluster types can use corosync, dlm, and gfs2. The main difference is the
management of clustered resources either by rgmanager(cman only) or pacemaker.
The cman plugin will capture data related to rgmanager.

In addition, there were duplicate commands that were removed and a couple of
commands and file captures that were added to different plugins that previously
was not in the cluster plugin.

Fixes: #587
